// admin/src/types/category.ts
export type TCategoryListItem = {
  categoryId: string;
  name: string;
  createdAt: string;
  updatedAt: string;
  createdBy: string;
  updatedBy: string;
}

export type TCategoryParams = {
  pageNo: number;
  pageSize: number;
};

export type TCategoryBody = {
  name: string;
}

export type TNewBookCategory = {
  categoryId: string;
  name: string
}